requirements.txt
#Heroku
Django>=1.6.1
gunicorn>=18.0
psycopg2==2.4.5
dj-database-url==0.2.0
dj-static
wsgiref==0.1.2
gevent>=0.13.8
South>=0.7.4
celery>=2.3.3
django-celery>=3.0.11
django-picklefield
newrelic

#Site Specific
django-redis-cache
django-taggit==0.10
django_compressor>=1.3
cssmin>=0.2.0
jsmin>=2.0.6
django-timezone-field
Markdown>=2.0.3

#caching
#johnny-cache

#Search
pyelasticsearch==0.6.1
django-haystack==2.1.0

#Storage
boto==2.9.6
django-queued-storage>=0.6
django-storages==1.1.8